requirements.txt
line-bot-sdk
flask
requests
openai